week07/requirements.txt
# Week 7 Buổi 13 — CDC to bronze (Kafka consumer → MinIO)
kafka-python>=2.0.2
minio>=7.2.0
